chore(dashboard): sandbox sheet cleanup
