+++ apps/dashboard/src/components/SandboxTable/types.ts
@@ -10,6 +10,7 @@ export interface SandboxTableProps {
   data: Sandbox[]
   sandboxIsLoading: Record<string, boolean>
   sandboxStateIsTransitioning: Record<string, boolean>
+  activeSandboxId?: string
   loading: boolean
   snapshots: SnapshotDto[]
   loadingSnapshots: boolean
