+++ apps/dashboard/src/components/sandboxes/SandboxDetailsSheet/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2025 Daytona Platforms Inc.
+ * SPDX-License-Identifier: AGPL-3.0
+ */
+
+export { default } from './SandboxDetailsSheet'
+export type { SandboxDetailsSheetProps, SandboxDetailsSheetTabValue } from './SandboxDetailsSheet'
