+++ apps/dashboard/src/components/ui/scroll-area.tsx
@@ -33,6 +33,7 @@ function ScrollArea({
   className,
   children,
   fade,
+  fadeOrientation,
   fadeSide,
   horizontal,
   vertical = true,
@@ -42,6 +43,7 @@ function ScrollArea({
 }: React.ComponentProps<typeof ScrollAreaPrimitive.Root> & {
   fade?: 'mask' | 'shadow'
   fadeOffset?: number
+  fadeOrientation?: 'vertical' | 'horizontal'
   fadeSide?: 'start' | 'end'
   horizontal?: boolean
   vertical?: boolean
@@ -49,7 +51,7 @@ function ScrollArea({
 }) {
   const rootRef = useRef<HTMLDivElement>(null)
   const viewportRef = useRef<HTMLDivElement>(null)
-  const fadeOrientation = horizontal ? 'horizontal' : 'vertical'
+  const resolvedFadeOrientation = fadeOrientation ?? (horizontal ? 'horizontal' : 'vertical')
   const setViewportRef = useCallback(
     (element: HTMLDivElement | null) => {
       viewportRef.current = element
@@ -89,13 +91,13 @@ function ScrollArea({
         'relative group/scroll-area',
         {
           'before:pointer-events-none before:absolute before:top-0 before:left-0 before:right-0 before:z-10 before:[height:var(--fade-offset)] before:bg-gradient-to-b dark:before:from-black/20 before:from-black/10 before:to-transparent before:transition-opacity before:duration-150 before:opacity-[min(1,calc(var(--offset-y-top)/20))]':
-            fade === 'shadow' && fadeOrientation === 'vertical' && fadeSide !== 'end',
+            fade === 'shadow' && resolvedFadeOrientation === 'vertical' && fadeSide !== 'end',
           'after:pointer-events-none after:absolute after:bottom-0 after:left-0 after:right-0 after:z-10 after:[height:var(--fade-offset)] after:bg-gradient-to-t dark:after:from-black/20 after:from-black/10 after:to-transparent after:transition-opacity after:duration-150 after:opacity-[min(1,calc(var(--offset-y-bottom)/20))]':
-            fade === 'shadow' && fadeOrientation === 'vertical' && fadeSide !== 'start',
+            fade === 'shadow' && resolvedFadeOrientation === 'vertical' && fadeSide !== 'start',
           'before:pointer-events-none before:absolute before:top-0 before:bottom-0 before:left-0 before:z-10 before:[width:var(--fade-offset)] before:bg-gradient-to-r dark:before:from-black/20 before:from-black/10 before:to-transparent before:transition-opacity before:duration-150 before:opacity-[min(1,calc(var(--offset-x-left)/20))]':
-            fade === 'shadow' && fadeOrientation === 'horizontal' && fadeSide !== 'end',
+            fade === 'shadow' && resolvedFadeOrientation === 'horizontal' && fadeSide !== 'end',
           'after:pointer-events-none after:absolute after:top-0 after:bottom-0 after:right-0 after:z-10 after:[width:var(--fade-offset)] after:bg-gradient-to-l dark:after:from-black/20 after:from-black/10 after:to-transparent after:transition-opacity after:duration-150 after:opacity-[min(1,calc(var(--offset-x-right)/20))]':
-            fade === 'shadow' && fadeOrientation === 'horizontal' && fadeSide !== 'start',
+            fade === 'shadow' && resolvedFadeOrientation === 'horizontal' && fadeSide !== 'start',
         },
         className,
       )}
@@ -112,17 +114,17 @@ function ScrollArea({
           {
             '!overflow-y-hidden': !vertical,
             '[mask-image:linear-gradient(to_bottom,transparent,black_min(var(--offset-y-top)*1px,var(--fade-offset)),black_calc(100%-min(var(--offset-y-bottom)*1px,var(--fade-offset))),transparent)]':
-              fade === 'mask' && fadeOrientation === 'vertical' && fadeSide === undefined,
+              fade === 'mask' && resolvedFadeOrientation === 'vertical' && fadeSide === undefined,
             '[mask-image:linear-gradient(to_bottom,transparent,black_min(var(--offset-y-top)*1px,var(--fade-offset)),black_100%)]':
-              fade === 'mask' && fadeOrientation === 'vertical' && fadeSide === 'start',
+              fade === 'mask' && resolvedFadeOrientation === 'vertical' && fadeSide === 'start',
             '[mask-image:linear-gradient(to_bottom,black_0,black_calc(100%-min(var(--offset-y-bottom)*1px,var(--fade-offset))),transparent)]':
-              fade === 'mask' && fadeOrientation === 'vertical' && fadeSide === 'end',
+              fade === 'mask' && resolvedFadeOrientation === 'vertical' && fadeSide === 'end',
             '[mask-image:linear-gradient(to_right,transparent,black_min(var(--offset-x-left)*1px,var(--fade-offset)),black_calc(100%-min(var(--offset-x-right)*1px,var(--fade-offset))),transparent)]':
-              fade === 'mask' && fadeOrientation === 'horizontal' && fadeSide === undefined,
+              fade === 'mask' && resolvedFadeOrientation === 'horizontal' && fadeSide === undefined,
             '[mask-image:linear-gradient(to_right,transparent,black_min(var(--offset-x-left)*1px,var(--fade-offset)),black_100%)]':
-              fade === 'mask' && fadeOrientation === 'horizontal' && fadeSide === 'start',
+              fade === 'mask' && resolvedFadeOrientation === 'horizontal' && fadeSide === 'start',
             '[mask-image:linear-gradient(to_right,black_0,black_calc(100%-min(var(--offset-x-right)*1px,var(--fade-offset))),transparent)]':
-              fade === 'mask' && fadeOrientation === 'horizontal' && fadeSide === 'end',
+              fade === 'mask' && resolvedFadeOrientation === 'horizontal' && fadeSide === 'end',
           },
         )}
       >
